+++ lib/class/artist.class.php
@@ -898,6 +898,10 @@ class Artist extends database_object implements library_item
                 $current_id = $artist_id;
                 Stats::migrate('artist', $this->id, $artist_id);
                 UserActivity::migrate('artist', $this->id, $artist_id);
+                Recommendation::migrate('artist', $this->id, $artist_id);
+                Share::migrate('artist', $this->id, $artist_id);
+                Shoutbox::migrate('artist', $this->id, $artist_id);
+                Tag::migrate('artist', $this->id, $artist_id);
                 Userflag::migrate('artist', $this->id, $artist_id);
                 Rating::migrate('artist', $this->id, $artist_id);
                 Art::migrate('artist', $this->id, $artist_id);