+++ modules/catalog/remote/remote.catalog.php
@@ -390,4 +390,4 @@ class Catalog_remote extends Catalog
 
         return null;
     }
-} // end of catalog class
+} // end of remote.catalog class